Programmers_월간 코드 챌린지 시즌1 - lv2_이진 변환 반복하기 : solved

+++ programmers/src/java_problemsolving/leveltwo/RepeatBinaryConversion.java
@@ -65,7 +65,7 @@ public class RepeatBinaryConversion {
             int postlength = s.length();
             answer[1] += prelength - postlength;
             int c = postlength;
-            int now = 0;
+            long now = 0;
             int idx = 0;
             while(c > 0) {
                 now += c%2*Math.pow(10, idx++);
